requirements.txt
streamlit>=1.32.0
google-generativeai>=0.4.0
python-dotenv>=1.0.0
plotly>=5.18.0
pandas>=2.0.0
